requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
line-bot-sdk==3.11.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
openai==1.51.2
